bird_interact_agent/requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.8.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.14.0
google-api-python-client==2.157.0
google-auth==2.23.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.61.0
grpcio==1.69.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.8.2
openai==1.55.3
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
requests==2.32.2
rsa==4.9
sniffio==1.3.1
socksio==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
datasets==3.2.0
psycopg2==2.9.9
PyMySQL==1.1.1
cryptography==3.4.8
pandas==2.1.0
numpy==1.26.1
sqlparse==0.5.2
rich==13.9.4
pymssql==2.3.2
jsonlines==4.0.0
config==0.5.1
gymnasium==1.1.1
docker==7.1.0
scikit-learn==1.7.0
tiktoken
sqlglot==26.25.3
google-cloud-aiplatform==1.92.0
google-auth==2.38.0
vertexai==1.71.1
